@@ -1,6 +1,6 @@
 {
   "name": "managed-timeout",
-  "version": "2.1.0",
+  "version": "3.0.0",
   "description": "setTimeout class wrapper with abilities to pause/resume, finish preemptively, and inspection",
   "repository": {
     "type": "git",
@@ -33,13 +33,14 @@
   ],
   "license": "ISC",
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.4.0",
-    "@typescript-eslint/parser": "^5.4.0",
-    "eslint": "^8.2.0",
-    "eslint-plugin-jest": "^25.2.4",
-    "jest": "^27.3.1",
-    "mockdate": "^3.0.5",
+    "@types/jest": "^29.2.6",
+    "@typescript-eslint/eslint-plugin": "^5.49.0",
+    "@typescript-eslint/parser": "^5.49.0",
+    "eslint": "^8.32.0",
+    "eslint-plugin-jest": "^27.2.1",
+    "jest": "^29.4.0",
+    "ts-jest": "^29.0.5",
     "tsup": "^6.5.0",
-    "typescript": "^4.5.2"
+    "typescript": "^4.9.4"
   }
 }
