@@ -1,3 +1,9 @@
+/** @type {import('ts-jest').JestConfigWithTsJest} */
 module.exports = {
-  timers: "fake",
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  fakeTimers: {
+      enableGlobally: true,
+  },
+  setupFilesAfterEnv: ['./tests/setup.ts']
 };
